Splits components and containers
